+++ src/constants/dummy.tsx
@@ -1,0 +1,18 @@
+import {Todo} from "../types/Todo"
+
+const mockTodo1:Todo = {
+    id: "1",
+    value: "やること1"
+};
+
+const mockTodo2:Todo = {
+    id: "2",
+    value: "やること2"
+}
+
+const mockTodo3:Todo = {
+    id: "3",
+    value: "やること3"
+}
+
+export const mockTodoList = [mockTodo1, mockTodo2, mockTodo3]
