+++ src/hooks/useTodo.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import { Todo, mockTodoList } from "../types/Todo"
+import { Todo } from "../types/Todo"
+import { mockTodoList } from "../constants/dummy"
 import "react-native-get-random-values";
 //import { v4 as uuidv4 } from "uuid";
