+++ src/types/Todo.ts
@@ -2,20 +2,3 @@ export type Todo = {
     id: string;
     value: string;
 };
-
-const mockTodo1:Todo = {
-    id: "1",
-    value: "やること1"
-};
-
-const mockTodo2:Todo = {
-    id: "2",
-    value: "やること2"
-}
-
-const mockTodo3:Todo = {
-    id: "3",
-    value: "やること3"
-}
-
-export const mockTodoList = [mockTodo1, mockTodo2, mockTodo3]
